LayoutTests/fast/speechsynthesis/speech-synthesis-cancel-queued-utterances-expected.txt

This tests that cancel() does not drop utterances queued after it is called. Regression test for https://bugs.webkit.org/show_bug.cgi?id=191745

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


u1: start
u1: error (canceled)
u2: start
Called cancel and queued u2, u3
u2: end
u3: start
u3: end
PASS u2Events.length is 2
PASS u2Events[0] is 'start'
PASS u2Events[1] is 'end'
PASS u3Events.length is 2
PASS u3Events[0] is 'start'
PASS u3Events[1] is 'end'
PASS successfullyParsed is true

TEST COMPLETE
